requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
aiolimiter>=1.1.0
python-docx>=0.8.11
pdfplumber>=0.10.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
dotenv